IR: Fix I64 incorrect format name

+++ src/main/java/org/systemf/compiler/ir/type/I64.java
@@ -12,7 +12,7 @@ public enum I64 implements IInteger {
 
 	@Override
 	public String getName() {
-		return "i32";
+		return "i64";
 	}
 
 	@Override
